backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.3.5
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
cors==1.0.1
python-multipart==0.0.6
